put <campaign-display> inside <bulbs-video-meta>
