+++ elements/bulbs-video/elements/meta/index.js
@@ -1,0 +1,2 @@
+import Meta from './meta';
+export default Meta;
